lambda_function/requirements.txt
tqdm==4.48.2
joblib==0.15.1
tensorflow==1.15.0
requests==2.18.4
pandas==0.23.4
six==1.11.0
boto3==1.17.27
nltk==3.5
numpy==1.19.5
pdf2image==1.14.0
botocore==1.20.27
pdf2jpg==1.0
Werkzeug==1.0.1
datefinder==0.7.1
opencv_contrib_python==4.2.0.34
Pillow==8.3.2